requirements.txt
mlflow
boto3
psycopg2-binary
python-dotenv
jupyter
papermill
pandas